+++ planner/stdlib.js
@@ -94,3 +94,56 @@ module.exports.objects_in_world = function(state) {
     }
     return list;
 };
+
+
+
+module.exports.test_satisfied = function(state, item, oneof, relation) {
+    // Find the object
+    var i = 0;
+    var j = -1;
+    for (var stack of state.stacks) {
+        j = stack.indexOf(item);
+        if (j !== -1) {
+            break;
+        }
+        i++;
+    }
+
+    switch (relation) {
+        case "holding":
+            var res = state.arms.some(function(arm) {
+                return arm.holding === item;
+            });
+            return res;
+
+        case "floor":
+            return j === 0;
+
+        case "inside":
+        case "ontop":
+            return j > 0 && oneof.contains(stack[j-1]);
+
+        case "above":
+            return j !== 1 && oneof.intersects(stack.slice(0, j));
+
+        case "under":
+            return j !== 1 && oneof.intersects(stack.slice(j+1));
+
+        case "beside":
+            if (j === -1) {
+                return false;
+            }
+            return  (i !== 0 && oneof.intersects(state.stacks[i-1])) ||
+                    (i !== state.stacks.length-1 && oneof.intersects(state.stacks[i+1]));
+
+        case "leftof":
+            return (j !== -1) && state.stacks.slice(i+1).flatten().intersects(oneof);
+
+        case "rightof":
+            return (j !== -1) && state.stacks.slice(0, i).flatten().intersects(oneof);
+
+        default:
+            throw "ERRORRRR: Planner does not know the relation: " + relation;
+    }
+
+};
